# Optimize build time with hard-source-webpack-plugin