<!-- Please don't delete this template -->

<!-- PULL REQUEST TEMPLATE -->
<!-- (Update "[ ]" to "[x]" to check a box) -->

**Summary**
Closes #444 #425 
<!-- If this PR fixes certain issues  -->
<!-- You can use the `Closes #issue_number` syntax here to close them -->

**What kind of change does this PR introduce?** (check at least one)

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update
- [ ] Refactor
- [ ] Docs
- [x] Build-related changes
- [ ] Other, please describe:

If changing the UI/CSS related code, please provide the **before/after** screenshot:

**Does this PR introduce a breaking change?** (check one)

- [ ] Yes
- [x] No

If yes, please describe the impact and migration path for existing applications:

**The PR fulfills these requirements:**

- [x] Briefly describing what this PR does in the title in [Angular Commit Message Conventions](https://git.io/fNGDG), e.g. `fix: rebuild when a page is added`

To avoid wasting your time, it's best to open a **feature request issue** first and wait for approval before working on it.

**Other information:**
The log
```
yarn run v1.19.0
$ node ../packages/saber/lib/cli.js
[info] Using config file: saber-config.js
[info] Using theme: ./src
[info] Using 9 plugins from config file
[hardsource:21992d5c] Using 10 MB of disk space.
[hardsource:21992d5c] Tracking node dependencies with: yarn.lock.
[hardsource:21992d5c] Reading from cache 21992d5c...
[hardsource:d4665d8a] Using 10 MB of disk space.
[hardsource:d4665d8a] Tracking node dependencies with: yarn.lock.
[hardsource:d4665d8a] Reading from cache d4665d8a...
[success] Compiled client successfully in 2.9s!
[info] Available at http://localhost:3000
[info] 73.2 MB memory used
[success] Compiled server successfully in 5.2s!
[info] (10:57:21) Rebuilding client due to changes in ~\issuehunt\saber\website\.saber\routes.js
[info] (10:57:21) Rebuilding server due to changes in ~\issuehunt\saber\website\.saber\routes.js
[hardsource:21992d5c] Using 10 MB of disk space.
[client] 62.50% building 7/8 modules 1 active ...style&index=0&id=2a183b29&scoped=true&lang=css&
[client] 94.50% after seal
[success] Compiled client successfully in 6.3s!
[info] Available at http://localhost:3000
[info] 141 MB memory used
[success] Compiled server successfully in 6.6s!
```